Decouple Navigator events from its results
Instead of extending the Navigator.Events interface
to add results or other events to the Navigator,
this commit decouples these two types of events.

Coupling these two causes a lot of unchecked casts
to be necessary and with that can lead to unexpected
crashes when a potential listener does not adhere to
the full contract.