Remove `CompositeDisposable.plusAssign`
The library includes a `CompositeDisposable.plusAssign(DisposableHandle)`
function, which is used to easily add `DisposableHandle`s to a
`CompositeDisposable`. However, this is rarely used but really gets in
the way when using RxJava's `plusAssign` in the sense that the
auto-import may suggest the wrong import.

The `DisposableHandle.asDisposable()` function should be enough to deal
with this easily.